Separates scoring and tracing for easier maintainability.
