+++ src/scaffold/langfuse_utils/__init__.py
@@ -1,0 +1,4 @@
+from .scoring import register_metric
+from .tracing import init_langfuse
+
+__all__ = ["init_langfuse", "register_metric"]
